refacted management code to support replication
JMX MBeans are now wrappers which use JBM core messages to invoke the management operations.
This way, management operations are properly invoked on both the live and backup nodes so that their state is correctly replicated.
Management attributes and operations w/o side-effect are called directly on the wrapped resource control.