Fix dead-lock due to taking multiple locks in inconsistent order.
startUpLock was used to control a small part of the start/stop cycle,
and was added before we had initialized2() synchronized. As the whole start/stop
cycle is already being controlled by synchronizing on this, namely:
- start();
- stop(bool) (most of it);
- initialize1()
- initialize2()

The conclusion is that we should just do away with 'startUpLock' whose name
could actually make people (incorrectly) conclude that it was (in practice)
controlling the start/stop cycle.