fix tck failure: com/sun/ts/tests/jms/core/topictests/TopicTests.java#consumerTests_from_ejb
Description: When clientID is preconfigured on CF, each connection from this CF will have same clientIDs,
which makes noLocal on topic subscriber's not working properly when you send a message from such a connection
and received it from another. Because of the two connections have same clientID, so the message won't get
received.

To solve this problem the connection's UID is used in creating topic subscribers so that differen connections
from the CF have different identifiers.