[HORNETQ-1172] Provide Websocket-like HTTP upgrade protocol
add http-upgrade-enable property (disabled by default) to
use HTTP Upgrade mechanism when connecting to the server.
The 1st request will be a HTTP Request + Upgrade: hornetq-remoting sent
to the server's Web container that'll reply with a HTTP 101 Switching
protocols.
When the client receives this response, it will switch to the
hornetq remoting protocol and remove the HTTP handler.