add netty property http-upgrade-endpoint
When using http-upgrade, the connector is multiplexed to the acceptor
through the Web server and its HTTP port
If there are multiple acceptors that are registered for the HTTP upgrade
mechanism, we need a HTTP header to match the connector initiating the
HTTP request with the acceptor that will take over the upgrade.
The property http-upgrade-endpoint is used for this case.