Re-arrange the example code to show the intent in a clearer manner.
While the example code is functional, it doesn't highlight the
dependencies between the lines as best it could.  The rearranged code
adds comments to help explain that producers write to 'addresses', that
'queues' are bound to an address and that client's read from a 'queue'.

Explaining this separation clearly is needed in order to understand
how to turn this example into a publish/subscribe model.