HORNETQ-1125 Added ServiceRegistry to Server
https://issues.jboss.org/browse/HORNETQ-1125

This patch adds an ServiceRegistry which stores references to
objects that are intended to be injected into HornetQ server.  This
allows WildFly to inject objects that Wildfly is responsible for
managing such as ExecutorServices. In addition Wildfly is able to create
objects in it's own class loader such as Intercepors and
ConnectorServiceFactories and inject them into HornetQ.  This means that
the HornetQ class loader does not need to be aware of the various
dependencies require to instantiate these objects.