HORNETQ-1419 Server doesn't deploy JMS Objects
When starting a standalone server using 'bin/hornetq run' the server
starts up but the jms objects such as Connection Factories and
Destinations configured in hornetq-jms.xml don't get deployed.

The reason behind this is that the "Run" command (Run.java) starts
the JMSServerManagerImpl without passing in a default config file and
we no longer put the default config dir into the server classpath
with the new startup method.

The fix is add a <jms> entry into hornetq.xml pointing to the default
hornetq-jms.xml file and in the "Run" command we pass the config
(if it configured) to JMSServerManagerImpl so that it will deploys
all the JMS stuff during start up.