BZ-1149936 HORNETQ-1427 Avoid orphaned consumers
https://issues.jboss.org/browse/HORNETQ-1427
https://bugzilla.redhat.com/show_bug.cgi?id=1149936

We have been reported about orphaned consumers cases.
The only cases where I saw this possibly happening was during a hypotethical
exception during server.close() or during a temporary blocking client (thread waiting to lock)

To avoid that I'm now removing connections at the end of loops and making the exception treatment
per consumer, so if any exception happened we will continue closing. at the worst case the connection
will still be on the maps and this won't be na issue any longer.