Escaped Characters in sender Id test
This is a test case that Justin Betram asked for. It is a replication of
some behavior I noticed where the sender ids (and other header fields)
that have characters that need to be escaped by the Stomp spec do not
appear to be. Here is the thread that started this:
https://developer.jboss.org/message/906432