Bug 1422335 Invalid id-cache-size in the default configuration
The default id-cache-size is 20000. It conflicts with the default
confirmation-window-size(1MB) used in ClusterConnection. When start
up a cluster the following WARN appears:

WARN  [org.hornetq.core.server] (Thread-16 (HornetQ-server-HornetQ
ServerImpl::serverUUID=dd22fac6-f327-11e6-a1c9-739b3b259544-27075288))
HQ222192: The size of duplicate cache detection (<id_cache-size/>)
appears to be too large. It should be no greater than the number
of messages that can be squeezed into conformation buffer
(<confirmation-window-size/>).

To fix it we adjust the confirmation-window-size to 10MB. Also
a test is added to guarantee it won't break this rule when this
default value is to be changed to any new value.

Merged from PR #2112