Bug1431406 specific properties not removed from messages
In a cluster if a message is routed to other nodes an internal
property (sth like _HQ_ROUTE_TOsfxxx) will be put to message
in order to route the message to its remote destination.
This property can be propogated to the consumers that
doesn't recognise the property.

For example if a 3rd party JMS consumer receives the message
it can get the property name by JMS api	Message.getPropertyNames()
and if the name is not confirm to some rules an erro will
occur.

The fix is that HorentQ message shouldn't expose such property
via JMS API.