HORNETQ-1551(2) Consumer hangs on corrupted 'forced delivery' message
It can happen when the consumer sends a 'forced delivery' then
waits forever while the connection is broken and the server's
'forced delivery' message never gets to consumer. If session is
reconnected, its consumer never knows and stays waiting.

To fix this, send a "forced delivery" message as part of session
reconnection.