HORNETQ-1578 Adding proper log message to  SharedNothingLiveActivation.isNodeIdUsed
When replicated live server starts up it needs to find out if a backup
is live in the cluster so it can decide whether to go live or perform
a fail back. The method SharedNothingLiveActivation#isNodeIdUsed() is
used in detecting a live backup. The method catches any exception and
return false. In the case that a live backup is there and the
DiscoveryGroup malfunctions (throwing exception), the restarting
live will wrongly think that no backup is live in cluster and go straight
live, resulting in 2 live servers at the same time. If that happens
no warning or debug messages is logged and therefore is hard to debug.