HORNETQ-1580 Large message's copy may be interfered by other threads
In LargeMessageImpl.copy(long) it need to open the underlying file in
order to read and copy bytes into the new copied message. However
there is a chance that another thread can come in and close the file
in the middle, making the copy failed with "channel is null" error.